# ARROW-9079: [C++] Write benchmark for arithmetic kernels