@wesm 2020-06-12
Thanks for working on this. I'll check the benchmarks on MSVC also