@wesm 2020-06-12
There don't seem to be issues on MSVC

https://gist.github.com/wesm/45be57393b2d9186f87faae228f12380/revisions