@wesm 2020-06-13
+1. I fixed a few lingering issues that jumped out at me, will merge this once build passes